+++ core-ng-api/src/main/java/core/framework/api/web/service/ResponseStatus.java
@@ -6,12 +6,13 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
 import static java.lang.annotation.ElementType.METHOD;
+import static java.lang.annotation.ElementType.TYPE;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
 /**
  * @author neo
  */
-@Target(METHOD)
+@Target({TYPE, METHOD})
 @Retention(RUNTIME)
 public @interface ResponseStatus {
     HTTPStatus value() default HTTPStatus.OK;