+++ core-ng/src/main/java/core/framework/api/validate/ValidationException.java
@@ -1,14 +1,17 @@
 package core.framework.api.validate;
 
+import core.framework.api.http.HTTPStatus;
 import core.framework.api.log.ErrorCode;
 import core.framework.api.log.Severity;
 import core.framework.api.util.Strings;
+import core.framework.api.web.service.ResponseStatus;
 
 import java.util.Map;
 
 /**
  * @author neo
  */
+@ResponseStatus(HTTPStatus.BAD_REQUEST)
 public class ValidationException extends RuntimeException implements ErrorCode {
     private static final long serialVersionUID = 9215299700445046388L;
     public final Map<String, String> errors;