+++ core-ng/src/main/java/core/framework/api/web/exception/TooManyRequestsException.java
@@ -1,11 +1,14 @@
 package core.framework.api.web.exception;
 
+import core.framework.api.http.HTTPStatus;
 import core.framework.api.log.ErrorCode;
 import core.framework.api.log.Severity;
+import core.framework.api.web.service.ResponseStatus;
 
 /**
  * @author neo
  */
+@ResponseStatus(HTTPStatus.TOO_MANY_REQUESTS)
 public final class TooManyRequestsException extends RuntimeException implements ErrorCode {
     private static final long serialVersionUID = 6657854760603154454L;