+++ core-ng/src/main/java/core/framework/impl/web/HTTPServerErrorHandler.java
@@ -6,17 +6,10 @@ import core.framework.api.log.ActionLogContext;
 import core.framework.api.log.ErrorCode;
 import core.framework.api.log.Severity;
 import core.framework.api.util.Exceptions;
-import core.framework.api.validate.ValidationException;
 import core.framework.api.web.ErrorHandler;
 import core.framework.api.web.Response;
 import core.framework.api.web.ResponseImpl;
-import core.framework.api.web.exception.BadRequestException;
-import core.framework.api.web.exception.ConflictException;
-import core.framework.api.web.exception.ForbiddenException;
-import core.framework.api.web.exception.MethodNotAllowedException;
-import core.framework.api.web.exception.NotFoundException;
-import core.framework.api.web.exception.TooManyRequestsException;
-import core.framework.api.web.exception.UnauthorizedException;
+import core.framework.api.web.service.ResponseStatus;
 import core.framework.impl.web.exception.ErrorResponse;
 import core.framework.impl.web.request.RequestImpl;
 import core.framework.impl.web.response.ResponseHandler;
@@ -45,10 +38,6 @@ public class HTTPServerErrorHandler {
             return;
         }
 
-        renderErrorPage(e, exchange, request);
-    }
-
-    private void renderErrorPage(Throwable e, HttpServerExchange exchange, RequestImpl request) {
         try {
             Response errorResponse = null;
             if (customErrorHandler != null) {
@@ -69,31 +58,20 @@ public class HTTPServerErrorHandler {
     }
 
     private Response defaultErrorResponse(Throwable e, String accept) {
-        HTTPStatus status;
-
-        if (e instanceof BadRequestException || e instanceof ValidationException) {
-            status = HTTPStatus.BAD_REQUEST;
-        } else if (e instanceof NotFoundException) {
-            status = HTTPStatus.NOT_FOUND;
-        } else if (e instanceof UnauthorizedException) {
-            status = HTTPStatus.UNAUTHORIZED;
-        } else if (e instanceof ForbiddenException) {
-            status = HTTPStatus.FORBIDDEN;
-        } else if (e instanceof MethodNotAllowedException) {
-            status = HTTPStatus.METHOD_NOT_ALLOWED;
-        } else if (e instanceof ConflictException) {
-            status = HTTPStatus.CONFLICT;
-        } else if (e instanceof TooManyRequestsException) {
-            status = HTTPStatus.TOO_MANY_REQUESTS;
-        } else {
-            status = HTTPStatus.INTERNAL_SERVER_ERROR;
-        }
+        HTTPStatus status = httpStatus(e);
 
         if (accept != null && accept.contains(ContentType.APPLICATION_JSON.mediaType())) {
             return Response.bean(errorResponse(e), status);
         } else {
             return Response.text(errorHTML(e), status, ContentType.TEXT_HTML);
         }
+    }
+
+    HTTPStatus httpStatus(Throwable e) {
+        ResponseStatus responseStatus = e.getClass().getAnnotation(ResponseStatus.class);
+        if (responseStatus != null) return responseStatus.value();
+
+        return HTTPStatus.INTERNAL_SERVER_ERROR;
     }
 
     private String errorHTML(Throwable e) {